moved rgb_hex_to_decimal() method to common/themes.py and done unit testing
